+++ ncarview/src/lib/libctrans/commondev.c
@@ -1,5 +1,5 @@
 /*
- *	$Id: commondev.c,v 1.16 1992-09-01 23:41:49 clyne Exp $
+ *	$Id: commondev.c,v 1.17 1992-09-09 15:09:02 clyne Exp $
  */
 #include <math.h>
 #include <stdio.h>
@@ -46,7 +46,7 @@ static	Ptype	*coordBuf = NULL;
 static	long	coordBufNum = 0;	/* the number of coords
 						currently in the buffer */
 static	long	coordBufSize = 0;       /* amount mem allocated         */
-static	unsigned long	maxPolyPoints = ~0;
+static	unsigned long	maxPolyPoints = ~(unsigned long) 0;
 
 int	commFillScaleFactor = 1;
 int	commHatchScaleFactor = 1;
@@ -262,7 +262,7 @@ CGMC *c;
 		 * last group. n = count of processed points .
 		 * p = count of processed unsent point specifications.
 		 */
-		while (n < (c->Pnum - 1)) {
+		while ((unsigned short) n < (unsigned short) (c->Pnum - 1)) {
 			if (Clipper(c->p[n].x,c->p[n].y,c->p[n+1].x,c->p[n+1].y,
 				&x1, &y1, &x2, &y2)) {
 
@@ -365,7 +365,7 @@ CGMC *c;
 	 * draw line segments one at a time
 	 */
 	if (ODD(c->Pnum)) c->Pnum--;    /* must be even */
-	for(i=0; i<c->Pnum; i+=2) {
+	for(i=0; (unsigned short) i<c->Pnum; i+=2) {
 		dev->line(c->p[i].x, c->p[i].y, c->p[i+1].x, c->p[i+1].y);
 	}
 	return(0);
@@ -410,7 +410,7 @@ int	fat_dot;
 
 	switch (MARKER_TYPE) {
 	case MARKER_X:	
-		for(i=0;i<c->Pnum;i++) {
+		for(i=0;(unsigned short) i<c->Pnum;i++) {
 			dev->line( c->p[i].x - offset, c->p[i].y - offset,
 			      c->p[i].x + offset, c->p[i].y + offset);
 
@@ -420,13 +420,13 @@ int	fat_dot;
 		break;
 
 	case MARKER_CIRCLE:
-		for(i=0;i<c->Pnum;i++) {
+		for(i=0;(unsigned short) i<c->Pnum;i++) {
 			quick_circle((int) c->p[i].x, (int) c->p[i].y, offset);	
 		}
 		break;
 
 	case MARKER_STAR:
-		for(i=0;i<c->Pnum;i++) {
+		for(i=0;(unsigned short) i<c->Pnum;i++) {
 			dev->line( c->p[i].x - offset, c->p[i].y,
 			      c->p[i].x + offset, c->p[i].y);
 
@@ -441,7 +441,7 @@ int	fat_dot;
 		}
 		break;
 	case MARKER_PLUS:
-		for(i=0;i<c->Pnum;i++) {
+		for(i=0;(unsigned short) i<c->Pnum;i++) {
 			dev->line( c->p[i].x - offset, c->p[i].y,
 			      c->p[i].x + offset, c->p[i].y);
 
@@ -450,7 +450,7 @@ int	fat_dot;
 		}
 		break;
 	case MARKER_DOT:
-		for(i=0;i<c->Pnum;i++) {
+		for(i=0;(unsigned short) i<c->Pnum;i++) {
 			/*
 			 * see if device needs a larger then normal dot
 			 */
@@ -591,7 +591,7 @@ CGMC *c;
 		}
 
 
-		for(i = 2; i < c->Pnum ; i++) {
+		for(i = 2; (unsigned short) i < c->Pnum ; i++) {
 			if (! Clipper(c->p[i-1].x,c->p[i-1].y,
 				c->p[i].x,c->p[i].y,
 				&x1,&y1,&x2,&y2)) {
